Start particle filter localization
modify pfilter module to compute weights directly from particle states rather than needing to create a hypothesis observation first. Also added dispersion to see how spread out the particles are indicating (hopefully) how confident we are in the estimate. The filter only uses x and y positions right now, and I am experimenting with the effect of changing the noise function and sigma. It seems lower sigma gives better results however the final weight_fn will not have access to the true value obviously, so sigma is not used. However it indicates we might want to be more harsh in the weights. Also, the noise parameters, resample factor or weight_fn can be adapted based on the confidence measurement for more robust localisation. The main step is to have observation be the distance/object measurements instead of the literal position.