Avoid obstacles that go out of view
Add "forced avoidance" which is triggered when we no longer are able to see an obstacle because we drove past it. When this is triggered, we enforce that the robot does not turn towards that direction until we have driven forward past the estimated distance of the obstacle when we last saw it. See todo_notes for more notes and current known fail cases/problems.